Remove dependency on internal files and use blazor package instead
