--- src/converters/piral-blazor/src/internal/globals.ts
@@ -1,5 +1,0 @@
-import '@dotnet/jsinterop';
-
-window['Blazor'] = {
-  _internal: {},
-};
